requirements.txt
# Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# Web Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
aiohttp>=3.9.0

# Database
motor>=3.3.2
pymongo>=4.6.0
dnspython>=2.4.2

# Auth & Sessions
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.2

# Utilities
requests>=2.31.0
icalendar>=5.0.11
schedule>=1.2.0

# Date/Time
python-dateutil>=2.8.2
